Removing duplicated loading skeletons and improve delete button conditions
