+++ app/NavBar.tsx
@@ -11,7 +11,7 @@ const NavBar = () => {
       label: 'Dashboard', href: '/'
     },
 
-    { label: 'Issues', href: '/issues' }
+    { label: 'Issues', href: '/issues/list' }
   ]
 
   const currentPath = usePathname()
