+++ app/issues/list/loading.tsx
@@ -1,16 +1,11 @@
+import { Table } from '@radix-ui/themes'
 import React from 'react'
-import { Button, Table } from '@radix-ui/themes'
-import Link from '../components/Link'
-import { prisma } from '@/prisma/client'
-import IssueStatusBadge from '../components/IssueStatusBadge'
-import delay from 'delay'
+import Skeleton from 'react-loading-skeleton'
+import 'react-loading-skeleton/dist/skeleton.css'
 import IssueActions from './IssueActions'
 
-
-const IssuesPage = async () => {
-
-  const issues = await prisma.issue.findMany();
-  await delay(2000)
+const LoadingIssuesPage = () => {
+  const issues = [1, 2, 3, 4, 5]
   return (
     <div>
       <IssueActions />
@@ -24,25 +19,23 @@ const IssuesPage = async () => {
         </Table.Header>
         <Table.Body>
           {issues.map(issue => (
-            <Table.Row key={issue.id}>
+            <Table.Row key={issue}>
               <Table.Cell>
-                <Link href={`/issues/${issue.id}`}>
-                  {issue.title}
-                </Link>
+                <Skeleton />
                 <div className='block md:hidden'>
-                  <IssueStatusBadge status={issue.status} />
+                  <Skeleton />
                 </div>
               </Table.Cell>
               <Table.Cell className='hidden md:table-cell'>
-                <IssueStatusBadge status={issue.status} />
+                <Skeleton />
               </Table.Cell>
-              <Table.Cell className='hidden md:table-cell'>{issue.createdAt.toDateString()}</Table.Cell>
+              <Table.Cell className='hidden md:table-cell'> <Skeleton /></Table.Cell>
             </Table.Row>
           ))}
         </Table.Body>
       </Table.Root>
-    </div >
+    </div>
   )
 }
 
-export default IssuesPage
+export default LoadingIssuesPage
