+++ app/issues/list/page.tsx
@@ -1,12 +1,16 @@
-import { Table } from '@radix-ui/themes'
 import React from 'react'
-import IssueStatusBadge from '../components/IssueStatusBadge'
-import Skeleton from 'react-loading-skeleton'
-import 'react-loading-skeleton/dist/skeleton.css'
+import { Table } from '@radix-ui/themes'
+import Link from '../../components/Link'
+import { prisma } from '@/prisma/client'
+import IssueStatusBadge from '../../components/IssueStatusBadge'
+import delay from 'delay'
 import IssueActions from './IssueActions'
 
-const LoadingIssuesPage = () => {
-  const issues = [1, 2, 3, 4, 5]
+
+const IssuesPage = async () => {
+
+  const issues = await prisma.issue.findMany();
+  await delay(2000)
   return (
     <div>
       <IssueActions />
@@ -20,23 +24,25 @@ const LoadingIssuesPage = () => {
         </Table.Header>
         <Table.Body>
           {issues.map(issue => (
-            <Table.Row key={issue}>
+            <Table.Row key={issue.id}>
               <Table.Cell>
-                <Skeleton />
+                <Link href={`/issues/${issue.id}`}>
+                  {issue.title}
+                </Link>
                 <div className='block md:hidden'>
-                  <Skeleton />
+                  <IssueStatusBadge status={issue.status} />
                 </div>
               </Table.Cell>
               <Table.Cell className='hidden md:table-cell'>
-                <Skeleton />
+                <IssueStatusBadge status={issue.status} />
               </Table.Cell>
-              <Table.Cell className='hidden md:table-cell'> <Skeleton /></Table.Cell>
+              <Table.Cell className='hidden md:table-cell'>{issue.createdAt.toDateString()}</Table.Cell>
             </Table.Row>
           ))}
         </Table.Body>
       </Table.Root>
-    </div>
+    </div >
   )
 }
 
-export default LoadingIssuesPage
+export default IssuesPage
